requirements.txt
ollama
whisper
ffmpeg-python
pyaudio
TTS
webrtcvad